// src/app/api/swagger/models/api-trade-dto-2.ts
export interface ApiTradeDto2 {
  AdditionalInfo?: string | null;
  AdditionalRequirements?: Array<ApiAdditionalRequirementDto2> | null;
  BusinessFlowType?: ApiTradeBusinessFlowEnum;
  CharacteristicFileGuid?: string;
  CharacteristicFromUi?: boolean | null;
  Characteristics?: Array<ApiCharacteristicDto> | null;
  ClientTimeZoneMinutesOffset?: number;
  ConditionsOfPayment?: string | null;
  ContactInfo?: ApiCustomerContactInfoAltDto | null;
  ContractAmountInfo?: string | null;
  CreatorUserId?: number | null;
  CurrencyCode?: string | null;
  CustomerId?: number;
  CustomerOrgFullName?: string | null;
  CustomerOrganizationId?: number;
  CustomerOrganizationInn?: string | null;
  CustomerOrganizationKpp?: string | null;
  CustomerWorkGroupNames?: Array<string> | null;
  DealConclusionTermInfo?: string | null;
  DealExecutionDate?: string | null;
  DeliveryPlace?: string | null;
  DeliveryTerms?: string | null;
  Description?: string | null;
  EisIntegrationData?: ApiEisIntegrationDataAltDto | null;
  EisTradeInfo?: ApiEisTradeInfoAltDto | null;
  ExternalFieldValues?: Array<ApiExternalFieldValueAltDto> | null;
  ExternalId?: string | null;
  ExternalSourcePlatformNumber?: string | null;
  ExternalSystemId?: number | null;
  FederalLawLimit?: ApiFederalLawLimitAltEnum | null;
  FillingApplicationEndDate?: string | null;
  FillingApplicationMinimumDateCountdown?: ApiFillingApplicationMinimumDateCountdownEnum | null;
  FillingApplicationsMinDays?: number | null;
  HasUncompletedEisOutgoingSaga?: boolean;
  Id?: number;
  InitialPrice?: number | null;
  InitialPriceWithVat?: number | null;
  InitialPriceWithoutVat?: number | null;
  InvitationIsAllRegions?: boolean;
  InvitationMethod?: ApiInvitationMethodEnum;
  InvitationRegions?: Array<string> | null;
  InvitationTags?: Array<string> | null;
  Invitations?: Array<ApiInvitationDto> | null;
  InvitedOrganizationIds?: Array<number> | null;
  IsAutoCompleteDeal?: boolean;
  IsBasedOnPublicOffer?: boolean | null;
  IsContractInElectronicForm?: boolean | null;
  IsCustomerAcceptByOkeiOffers?: boolean;
  IsCustomerAcceptOnlyTop5Offers?: boolean;
  IsDifferenceProtocolSendOnce?: boolean;
  IsFederalLawsEnable?: boolean;
  IsImmediate?: boolean;
  IsImportPhaseout?: boolean | null;
  IsInitialPriceDefined?: boolean;
  IsInitialPriceReduction?: boolean;
  IsLotPriceEnteredWithVAT?: boolean;
  /** @deprecated */IsLotPriceWithVat?: boolean;
  IsManualEnterPriceForEachProduct?: boolean;
  IsModificationSentToOos?: boolean;
  IsOnlySmp?: boolean;
  IsParentContractNumberNotSpecified?: boolean | null;
  IsPositionalVatEnable?: boolean;
  IsPriceCorridorUsed?: boolean;
  IsRetradingEnabled?: boolean;
  IsRussianPurchaseObjectHasAdvantage?: boolean;
  IsSanctionedPurchase?: boolean | null;
  IsSendToEis?: boolean;
  IsSingleWinner?: boolean | null;
  IsUnitBidding?: boolean;
  KeepGuaranteeForAllParticipants?: boolean;
  KeepGuaranteeForParticipantsCount?: number | null;
  KladrRegionCode?: string | null;
  KladrRegionCodes?: Array<ApiDictionaryItemAltDto> | null;
  LastModificationDate?: string | null;
  LotItemsSumPrice?: number | null;
  MaxDealPrice?: number | null;
  MaxDealPriceSumVat?: number | null;
  MaxDealPriceWithVat?: number | null;
  MaxDealPriceWithoutVat?: number | null;
  ModificationInfo?: string | null;
  OfferId?: number | null;
  OrganizerId?: number;
  ParentContractNumber?: string | null;
  ParentNotTookPlaceTradeId?: number | null;
  PlanId?: number | null;
  PlanNumber?: string | null;
  PlanPositionNumber?: string | null;
  PlanStartingPrice?: number | null;
  PlanStartingPriceWithoutVat?: number | null;
  PlanedDealSignDate?: string | null;
  Products?: Array<ApiProductDto2> | null;
  RateVat?: ApiAvailableVatTypeEnum;
  RowVersion?: number;
  SelectOffersOnlyFromProposedBySuppliers?: boolean;
  SourcePlatform?: ApiSourcePlatformEnum;
  Specifications?: Array<ApiTradeLotSpecificationDto> | null;
  SupplierWithSmallBusinessRequirements?: boolean;
  TenantKey?: number;
  TradeDocuments?: Array<ApiTradeDocumentDto> | null;
  TradeIdentificationCode?: string | null;
  TradeLotIds?: Array<number> | null;
  TradeName?: string | null;
  TradeState?: ApiTradeStateEnum;
  VatRateState?: ApiVatRateStateEnum;
  WorkGroupIds?: Array<number> | null;
  ZmoFinanceSource?: ApiZmoFinanceSourceEnum | null;
  ZmoFzType?: number | null;
}